Add design components, theme tokens & fixes
Introduce a set of reusable UI components and theme tokens to align the app with the FoodieNow design system: added components (EmptyState, FoodCard, PriceTag, QuantitySelector, RatingBadge, RestaurantCard, StatusChip, VoucherCard), expanded ShimmerEffect with multiple skeletons and a 1200ms shimmer animation, and updated shapes/spacing/typography/elevation/color tokens (Spacing, Elevation, ShapeImage, additional semantic colors and text styles). Also register MockVoucherRepository in RepositoryModule and add a VoucherRepository interface. Fix navigation argument type for PaymentResult.amount from Float to Long. These changes centralize design tokens and common UI pieces to improve consistency and reuse across screens.

+++ app/src/main/java/com/example/foodienow/core/navigation/Screen.kt
@@ -15,7 +15,7 @@ sealed class Screen(val route: String) {
     object Cart : Screen("cart_screen")
     object Payment : Screen("payment_screen")
     object PaymentResult : Screen("payment_result_screen/{orderId}/{amount}/{methodLabel}") {
-        fun createRoute(orderId: String, amount: Double, methodLabel: String) =
+        fun createRoute(orderId: String, amount: Long, methodLabel: String) =
             "payment_result_screen/$orderId/$amount/${android.net.Uri.encode(methodLabel)}"
     }
     object Profile : Screen("profile_screen")
