+++ app/src/main/java/com/example/foodienow/feature/customer_home/components/CurrencyExt.kt
@@ -3,7 +3,7 @@ import java.text.DecimalFormat
 import java.text.DecimalFormatSymbols
 import java.util.Locale
 
-fun Double.formatPrice(): String {
+fun Long.formatPrice(): String {
     val symbols = DecimalFormatSymbols(Locale("vi", "VN"))
     symbols.groupingSeparator = '.'
     val decimalFormat = DecimalFormat("###,###", symbols)
